Fixed MITM security vulnerability
Previously OpenSSL connections were not verified.  This allowed anyone
along the path of the request to perform a MITM attack on the connection
undetected and snoop content, change content or acquire login
credentials which may lead to further data disclosure.

Now the OpenSSL default certificate store is used to verify connections.
The connection will not succeed if the certificate is not trusted.